Refactor: 회원가입 로직을 사용자 프로필 및 검증 모듈로 분리
- 회원가입 관련 URL을 user.profile 및 user.validate로 이동
