+++ config/urls.py
@@ -21,5 +21,6 @@ from django.urls import path, include
 urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/", include("core.auth.login.urls")),
-    path("api/", include("core.auth.signup.urls")),
+    path("api/", include("core.user.profile.urls")),
+    path("api/", include("core.user.validate.urls")),
 ]
