+++ core/user/validate/urls.py
@@ -1,20 +1,18 @@
 from django.urls import path
 from .views import (
-    SignUpView,
     UsernameValidationView,
     BojUsernameValidationView,
 )
 
 urlpatterns = [
-    path("auth/signup/", SignUpView.as_view(), name="signup"),
     path(
-        "auth/signup/validate/username/",
+        "user/validate/username/",
         UsernameValidationView.as_view(),
-        name="signup_validate_username",
+        name="validate_username",
     ),
     path(
-        "auth/signup/validate/boj/",
+        "user/validate/boj/",
         BojUsernameValidationView.as_view(),
-        name="signup_validate_boj",
+        name="validate_boj_username",
     ),
 ]
