+++ core/user/validate/views.py
@@ -3,30 +3,9 @@ from rest_framework.response import Response
 from rest_framework import status
 from rest_framework.permissions import IsAuthenticated
 from .serializers import (
-    UserSignUpSerializer,
     UsernameValidationSerializer,
     BojUsernameValidationSerializer,
 )
-
-
-class SignUpView(APIView):
-    permission_classes = [IsAuthenticated]
-
-    def post(self, request):
-        user = request.user
-
-        serializer = UserSignUpSerializer(
-            user, data=request.data, context={"request": request}, partial=True
-        )
-
-        if serializer.is_valid():
-            serializer.save()
-            return Response(
-                {"message": "회원가입이 완료되었습니다.", "user": serializer.data},
-                status=status.HTTP_200_OK,
-            )
-
-        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 class UsernameValidationView(APIView):
