+++ core/user/validators.py
@@ -1,0 +1,34 @@
+import requests
+from rest_framework import serializers
+
+from core.models import User
+
+
+def validate_username_for_user(user, value: str) -> str:
+    if User.objects.filter(username=value).exclude(id=user.id).exists():
+        raise serializers.ValidationError("이미 존재하는 닉네임입니다.")
+
+    return value
+
+
+def validate_boj_username_for_user(user, value: str) -> str:
+    if User.objects.filter(boj_username=value).exclude(id=user.id).exists():
+        raise serializers.ValidationError("이미 존재하는 백준 계정입니다.")
+
+    try:
+        response = requests.get(
+            f"https://solved.ac/api/v3/user/show?handle={value}",
+            timeout=5,
+        )
+        if response.status_code == 404:
+            raise serializers.ValidationError("존재하지 않는 백준 계정입니다.")
+        if response.status_code != 200:
+            raise serializers.ValidationError(
+                "백준 계정 확인 중 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
+            )
+    except requests.RequestException:
+        raise serializers.ValidationError(
+            "백준 계정 확인 중 오류가 발생했습니다. 잠시 후 다시 시도해주세요."
+        )
+
+    return value
